Move canny to skimage2
