+++ src/skimage2/feature/__init__.pyi
@@ -1,5 +1,7 @@
 __all__ = [
+    "canny",
     "peak_local_max",
 ]
 
+from ._canny import canny
 from ._peaks import peak_local_max
