+++ tests/skimage2/feature/test_canny.py
@@ -3,8 +3,9 @@ import numpy as np
 import pytest
 from skimage._shared.testing import assert_equal
 from scipy.ndimage import binary_dilation, binary_erosion
-from skimage import data, feature
+from skimage import data
 from skimage.util import img_as_float
+from skimage2 import feature
 
 
 class TestCanny(unittest.TestCase):
